♻️ Refacto fetch user's groups and relations
