+++ pages/groupe/creation.vue
@@ -5,6 +5,12 @@
 </template>
 
 <script setup lang="ts">
+const { fetchUserGroupsAndRelations } = groupHook()
+
+onMounted(async () => {
+  await fetchUserGroupsAndRelations()
+})
+
 definePageMeta({
   layout: 'auth',
   isAuth: true,
