+++ pages/groupe/index.vue
@@ -24,35 +24,13 @@
 </template>
 
 <script setup lang="ts">
-import { uniq } from '@antfu/utils'
-import {
-  useEmployeeStore,
-  useGroupStore,
-  useUiStore,
-} from '~~/store'
+import { useGroupStore } from '~~/store'
 
-const { IncLoading, DecLoading } = useUiStore()
-const { fetchByUser } = groupHook()
-const { fetchMany } = employeeHook()
-
+const { fetchUserGroupsAndRelations } = groupHook()
 const groupStore = useGroupStore()
-const employeeStore = useEmployeeStore()
 
 onMounted(async () => {
-  IncLoading()
-  if (groupStore.getIsEmpty) {
-    await fetchByUser()
-  }
-
-  const missingEmployees = uniq(groupStore.getAllArray
-    .reduce((acc, emp) => [...acc, ...emp.employeeIds], [] as number[]))
-    .filter(id => !employeeStore.isAlreadyInStore(id))
-
-  if (missingEmployees.length > 0) {
-    await fetchMany(missingEmployees)
-  }
-
-  DecLoading()
+  await fetchUserGroupsAndRelations()
 })
 
 definePageMeta({
